+++ CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 
+## [Unreleased]
+### Fixed
+- Added upper bound to number of documents to retrieve in `BM25Retriever`.
+
+
 ## [v4.1.0] - 2024-11-05
 ### Added
 - Added a new `RaggerPipeline` class which allows easier integration with Open WebUI.